requirements.txt
PySide6>=6.7.0
PyMuPDF>=1.24.0
Pillow>=10.0.0
pytesseract>=0.3.10
numpy>=1.26.0
opencv-python>=4.10.0